Improve TSIP packet sync
If we get an unexpected DLE sequence mid-packet, we're never going
to decode this packet correctly. Before, we were resetting the
decoder state to wait for the next DLE to be beginning-of-packet, but
it's possible to get permanently out-of-sync that way. Instead, treat
the DLE we *just* read as the first DLE of a new packet, providing
another opportunity to get into sync.