Yet more big changes to timing
FLL is rewritten yet again; it's simpler and I think on a good
theoretical basis at this point. A sort of recursive low-pass filter is
put on the front end of all of the timing to keep the 32ns jumps from
causing so much dancing around.

It's been running this way for months and quite stable, I'm just
getting around to committing it now.